refact(py): Refactor structure (#481)
* 目录结构重构
* update config
* rename dns to provider
* patch version info
* remove version replace
* add ENV GITHUB_REF_NAME to docker build
---------
