+++ ddns/__init__.py
@@ -1,0 +1,18 @@
+# -*- coding:utf-8 -*-
+"""
+ddns Package
+"""
+
+
+# __version__ = "${BUILD_VERSION}@${BUILD_DATE}"  # CI 时会被Tag替换
+__version__ = "0.0.0"
+__description__ = "automatically update DNS records to my IP [域名自动指向本机IP]"
+
+__doc__ = """
+ddns[%s]
+(i) homepage or docs [文档主页]: https://ddns.newfuture.cc/
+(?) issues or bugs [问题和帮助]: https://github.com/NewFuture/DDNS/issues
+Copyright (c) New Future (MIT License)
+""" % (
+    __version__
+)
